bad type, int->size_t

+++ crypt_blowfish.c
@@ -776,7 +776,7 @@ static char *BF_crypt(const char *key, const char *setting,
 	return output;
 }
 
-int _crypt_output_magic(const char *setting, char *output, int size)
+int _crypt_output_magic(const char *setting, char *output, size_t size)
 {
 	if (size < 3)
 		return -1;
@@ -812,7 +812,7 @@ int _crypt_output_magic(const char *setting, char *output, int size)
  * setting.
  */
 char *_crypt_blowfish_rn(const char *key, const char *setting,
-	char *output, int size)
+	char *output, size_t size)
 {
 	const char *test_key = "8b \xd0\xc1\xd2\xcf\xcc\xd8";
 	const char *test_setting = "$2a$00$abcdefghijklmnopqrstuu";
@@ -874,7 +874,7 @@ char *_crypt_blowfish_rn(const char *key, const char *setting,
 }
 
 char *_crypt_gensalt_blowfish_rn(const char *prefix, unsigned long count,
-	const char *input, int size, char *output, int output_size)
+	const char *input, int size, char *output, size_t output_size)
 {
 	if (size < 16 || output_size < 7 + 22 + 1 ||
 	    (count && (count < 4 || count > 31)) ||
