+++ crypt_blowfish.h
@@ -17,11 +17,13 @@
 #ifndef _CRYPT_BLOWFISH_H
 #define _CRYPT_BLOWFISH_H
 
-extern int _crypt_output_magic(const char *setting, char *output, int size);
+#include <stddef.h>
+
+extern int _crypt_output_magic(const char *setting, char *output, size_t size);
 extern char *_crypt_blowfish_rn(const char *key, const char *setting,
-	char *output, int size);
+	char *output, size_t size);
 extern char *_crypt_gensalt_blowfish_rn(const char *prefix,
 	unsigned long count,
-	const char *input, int size, char *output, int output_size);
+	const char *input, int size, char *output, size_t output_size);
 
 #endif
